Created `utils.py` to house utility functions
`read_json` takes a path and attempts to output a loaded Bitwarden vault or TOTP object/dictionary

`write_json` takes a path and output data and attempts to write the accounts not using 2FA to the specified file

`validate_bitwarden_vault` takes a loaded Bitwarden vault and attempts to validate it, throwing errors if the structure doesn't align with what is required for usage

`get_top_level_domain` takes a URI and returns only the top-level domain (e.g., example.com) to support accurately finding matches in `totp_json`

`get_accounts_with_unused_2fa` takes a Bitwarden vault and TOTP object/dictionary and iterates through each account and each uri inside that account to identify matches with `totp_json`. If a match is found and TOTP is not being used, that account is added to a running list of accounts with name, uri, and optional documenation link. When complete, a list of supported accounts with unused 2FA is returned